+++ db/migrate/20210627112659_create_feedbacks.rb
@@ -2,7 +2,7 @@ class CreateFeedbacks < ActiveRecord::Migration[6.1]
   def change
     create_table :feedbacks do |t|
       t.string :text
-      t.numeric :user_id
+      t.bigint :user_id
       t.numeric :starts
 
       t.timestamps
