+++ db/migrate/20210627113734_create_room_services.rb
@@ -1,8 +1,8 @@
 class CreateRoomServices < ActiveRecord::Migration[6.1]
   def change
     create_table :room_services do |t|
-      t.numeric :room_id
-      t.numeric :service_id
+      t.bigint :room_id
+      t.bigint :service_id
 
       t.timestamps
     end
