+++ db/migrate/20210627115526_create_issues.rb
@@ -1,8 +1,8 @@
 class CreateIssues < ActiveRecord::Migration[6.1]
   def change
     create_table :issues do |t|
-      t.numeric :category_id
-      t.numeric :user_id
+      t.bigint :category_id
+      t.bigint :user_id
       t.string :text
       t.string :status
