+++ db/migrate/20210709074224_create_room_images_table.rb
@@ -1,8 +1,8 @@
 class CreateRoomImagesTable < ActiveRecord::Migration[6.1]
   def change
     create_table :room_images do |t|
-      t.numeric :room_id
-      t.numeric :image_id
+      t.bigint :room_id
+      t.bigint :image_id
 
       t.timestamps
     end
